Refactor actions that are part of api into api folder for new position form
